Refactor prom-metrics datalog loop to avoid drift
This is a refactoring of the datalog loop for prometheus metrics
to avoid drift by marching the time interval using a calculation
from the initial interval instead of using a new time stamp to
start the new interval.

It calculates the average interval duration and keeps track of
the number of times drift occurs, reporting that in the warning log
message when drift occurs.